init RCPSP implementation
